Core: Tasks subsystem rework.
Tasks subsystem was reworked to use deboost.context instead of windows fibers or ucontext/setjmp mix. New library takes proper care to update TIB on windows therefore there is no need to keep using windows implementation. Code is simplified as a result.

Tasks no longer use a timer for sleeping. Instead they save time task is supposed to run next. This allowed task scheduler to sort tasks by this time value. This optimization will be very useful with huge task count. Hitting first task that is not scheduled to run allows early exit of the scheduler loop without a need to check all remaining tasks which are sleeping.

Task container in the scheduler was changed from linked list to a vector. Linked list was used due to cheap removal of items from the middle of the list, because nature of the tasks is such that their lifetime may vastly differ. Dead tasks are not removed from vector, instead they are replaced with a null pointer. Sorting of tasks mentioned earlier allows another optimization - pushing null task pointers to the end of the vector and simply trimming the vector to get rid of these null pointers.

Valgrind support for task stacks was added.

Default stack size was increased from 5K to 64K.

Use of `thread_local` was reduced to bare minimum and there is define to disable it completely.