Editor: Usability improvements when right-clicking things in scene view and node tree.
It is now possible to right-click object in the viewport. It opens context menu of right-clicked node. Other nodes can be clicked while context menu is open as well.

In node tree it is now possible to right-click nodes while context menu of other node is open. This saves extra pointless clicks that were needed in order to get rid of previous node context menu.