CSharp: Another huge rework - detached wrapper generation from c++ ast.
Every problem can be solved with one more layer of abstraction, right? C++ ast is parsed into a tree-like structure of declarations. This tree then is traversed in order to generate wrapper code. This paves a way for altering wrapper code output by modifying this declarations tree. For example we will now be able to move constants from global scope to a dummy class in order to have them appear in c# class.