Remove call to SetupInstance() in constructor from native pointer.
Constructing from native pointer implies that object is already set up. Calling SetupInstance() could potentially break the state of the object. For example if subclass of `Component` is created by user or by factory it will always execute constructor taking `Context` parameter. This constructor calls SetupInstance(). Managed object is tied to a lifetime of native object so in general we can assume that constructor `Component`'s subclass taking IntPtr will never be executed. When instance of class that does not support inheritance is returned from native code and constructor taking IntPtr is invoked it is OK to not call setup method because there is no setup to be done anyway.