ThirdParty/Scene/Editor: Remove yaml-cpp dependency and yaml-related functions.
YAML looks good on the surface, however it is a bad choice for serialization. Problem with YAML is that it is too simple and too complex at the same time. Complexity is less of an issue because automatic serialization does not care. However simplicity becomes really problematic when conflicts happen. For example if two users edited same scene in around same area git merge would likely fail with a conflict that has to be resolved manually. YAML being simple this would present user with a very long and confusing columns of text that have to be resolved manually. Judging where scope spans is problematic. In addition to this people are usually less familiar with YAML than with XML or JSON therefore it creates a bit higher barrier of entry for new users. Due to these reasons YAML was removed from the engine.

Any serializable assets will be saved as XML from now on. XML is not a perfect format, however few of it's advantages in big merges swayed the decision. In XML end of the scope is very explicit and exposes name of the scope. This will definitely be helpful in big merges. JSON ends scopes with plain } which is not much more helpful than space-aware scoping in YAML. Another advantage of XML is attributes, which allows to have a more compact file. Entire variant value can be described in one line including it's name/type/value. JSON would require a dictionary with multiple keys which would span at least 4 lines (and more in some cases).

Various settings files that will be serialized in JSON. This format is a bit easier and natural to serialize using engine classes. Said files should not grow large therefore editing them manually is a tiny bit user-friendlier than XML. Some of these files are not meant to be edited manually at all.