Core: Add a private "end frame" event for use by tools.
Plugin handling code must run after all user code at the end of the frame. Subscribing to existing E_ENDFRAME event is not good enough because some plugin code may be subscribed to this event and run right after reloading. This would create situation where plugin gets to handle end of frame that was never started as far as plugin is concerned.