Editor: Make right clicks not register on scene view when mouse was in a hidden state on this frame.
This dual check is important to make it work independent of render order. For example if inspector renders after scene tab then ui::GetMouseCursor() may report a cursor that is not ImGuiMouseCursor_None because cursor is reset on each frame. However if inspector renders before scene tab then mouse is still in a hidden state and click should not register.