AABB/OBB raycast to return box normal when possible (#505)
Ray can now be tested against BoundingBox and return hit normal along
with distance. This normal could be use to populate raycast result where
possible.

A sample scene added to illustrate how to use the raycast. The index 99
of the sample reflects that it isn't a sample for new functionality, it
is just an illustration on how existing functionality works.

Co-authored-by: Eugene Kozlov <nekoeuge@gmail.com>