Skymesh and Skyplane shaders (#638)
Skymesh: a new shader to replace Skydome. It renders any mesh as a
skybox and keeps UV mapping of original mesh.
Skyplane: renders an image on background
  Alignment defines how to align image. 0.5, 0.5 for center.
PixelPerfect when set to 1 the image is rendered as it's original scale
pixel-to-texel. When set to 0 it scales to cover the entire viewport.

https://youtu.be/02xiX1vkBes